// package/billing/payment/core/interface/dunning-schedule.interface.ts
import { DunningStage } from '../enum/dunning-stage.enum';
import { PaymentStatus } from '../../../payment/core/enum/payment-status.enum';

export interface DunningAttemptResult {
  success: boolean;
  status: PaymentStatus;
  nextAttemptScheduled: boolean;
  nextAttemptAt?: Date;
  stage: DunningStage;
}

export interface DunningSchedule {
  stage: DunningStage;
  daysFromFirstFailure: number;
  actions: string[];
}